Weather App/pyproject.toml
[tool.poetry]
name = "Weather App"
version = "0.1.0"
description = "This application will generate weather information based on user input for the current day"
authors = ["Ian Maze <mazeian93@outlook.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
streamlit = "^1.17.0"
load-dotenv = "^0.1.0"
requests = "^2.28.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
